Add to `options` object instead of a new parameter
